feature : 정량 평가 질문 클러스터 별로 가져오는 api 구현

+++ src/main/java/com/kcc/fillin/statistic/dto/ClusterAnswerDto.java
@@ -17,7 +17,7 @@ import lombok.ToString;
 @ToString
 public class ClusterAnswerDto {
 	// private Long answerSeq;
-	private Long answerOrderNum;
+	private Integer answerOrderNum;
 	private String answerName;
 	private List<Long> clusterList;
 	// private List<ClusterDto> clusterList;
